feat (tests): Some fixes trying to move out some TNS import sentences which are not supported by jest.
