+++ platform/nativescript/runtime/index.js
@@ -1,4 +1,3 @@
-import { run, on, launchEvent } from 'tns-core-modules/application'
 import { warn } from 'core/util/index'
 import { patch } from './patch'
 import { mountComponent } from 'core/instance/lifecycle'
@@ -54,6 +53,8 @@ Vue.prototype.$mount = function(el, hydrating) {
 }
 
 Vue.prototype.$start = function() {
+  const application = require('tns-core-modules/application')
+
   let self = this
   const AppConstructor = Vue.extend(this.$options)
 
@@ -62,7 +63,7 @@ Vue.prototype.$start = function() {
     Vue.component(entry.meta.component.name, entry.meta.component)
   })
 
-  on(launchEvent, args => {
+  application.on(application.launchEvent, args => {
     if (self.$el) {
       self.$destroy()
       self = new AppConstructor()
@@ -72,7 +73,7 @@ Vue.prototype.$start = function() {
     args.root = self.$el.nativeView
   })
 
-  run()
+  application.run()
 }
 
 // Define a `nativeView` getter in every NS vue instance
